LayoutTests/js/array-with-double-push-expected.txt

This test checks for regression against 117281: Crash in V8 benchmarks set in ARM,softfp,EABI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is 5000000000
PASS successfullyParsed is true

TEST COMPLETE
